Patch version placeholder in teliod
Libtelio has a version placeholder which when promoting
the build is then patched(string is replaced in binary).

As a result, when building "teliod" binary it contains
the placeholder of the version.

In order to interfere with libtelio as little as possible, the version
patching is done explicitly in here, OpenWRT Makefile.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>